requirements.txt
pymongo==4.6.0
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
numpy==1.26.0
scikit-learn==1.3.0
nltk==3.8.0
speech-recognition==3.10.0
pydub==0.25.1
python-dotenv==1.0.0
cohere==5.0.0
anthropic==0.28.0
transformers==4.41.2
huggingface_hub==0.22.2
fastapi==0.111.0
uvicorn==0.30.1
